Refactored monster type based calculations (not sharing a method through usage of an Enum), to reduce checks to some degree
