--- src/monsterTyping/MonsterTypeModifierGroup.ts
@@ -1,11 +1,0 @@
-/** Groups of modifiers */
-export enum MonsterTypeModifierGroup {
-    DamagePercent = "DamagePercent",
-    DamageTakenPercent = "DamageTakenPercent",
-    MaxHitPercent = "MaxHitPercent",
-    MaxHitFlat = "MaxHitFlat",
-    MinHitBasedOnMaxHit = "MinHitBasedOnMaxHit",
-    FlatMinHit = "FlatMinHit",
-    GlobalAccuracy = "GlobalAccuracy",
-    DamageReduction = "DamageReduction"
-}
